Pembudidaya Nelayan - Validasi

+++ app/Http/Controllers/App/NelayanController.php
@@ -40,7 +40,6 @@ class NelayanController extends Controller
 		$this->validate($r,[
 				'nik' => 'required|unique:users',
 				'no_kartu_nelayan' => 'required|unique:users',
-				'id_sarana' => 'required',
 			]);
 		
 		// Validasi Jabatan
@@ -57,7 +56,7 @@ class NelayanController extends Controller
 
 			$r->session()->flash('gagal','GAGAL!!! Jabatan <b>'.$vj->nama_jabatan.'</b> pada kelompok <b>'.$vj->nama_kelompok.'</b> telah ada');
 
-			return redirect(route('nelayan'));
+			return redirect(route('nelayan'))->withInput();
 			exit;
 		}
 
@@ -79,13 +78,15 @@ class NelayanController extends Controller
 		$id = $pb->id;
 
 		// Simpan sarana
-		foreach ( $r->id_sarana as $val ){
-			$record['id_sarana']  = $val; 
-			$record['id_user']    = $id; 
-			$records[] = $record;
-		}
+		if ( $r->id_sarana ) {
+			foreach ( $r->id_sarana as $val ){
+				$record['id_sarana']  = $val; 
+				$record['id_user']    = $id; 
+				$records[] = $record;
+			}
 
-		DB::table('app_kepemilikan_sarana')->insert( $records );
+			DB::table('app_kepemilikan_sarana')->insert( $records );
+		}
 
 		$r->session()->flash('success','Data tersimpan');
 
@@ -106,7 +107,6 @@ class NelayanController extends Controller
 		$this->validate($r,[
 				'nik' => 'required|unique:users,id,'.$r->id,
 				'no_kartu_nelayan' => 'required|unique:users,id,'.$r->id,
-				'id_sarana' => 'required',
 			]);
 
 		// Validasi Jabatan
@@ -144,16 +144,20 @@ class NelayanController extends Controller
 
 		$id = $r->id;
 
-		// Hapus lalu simpan kembali
-		KepemilikanSarana::where('id_user', $id)->delete();
 
-		foreach ( $r->id_sarana as $val ){
-			$record['id_sarana']  = $val; 
-			$record['id_user']    = $id; 
-			$records[] = $record;
+		if ( $r->id_sarana ) {
+
+			// Hapus lalu simpan kembali
+			KepemilikanSarana::where('id_user', $id)->delete();
+
+			foreach ( $r->id_sarana as $val ){
+				$record['id_sarana']  = $val; 
+				$record['id_user']    = $id; 
+				$records[] = $record;
+			}
+
+			DB::table('app_kepemilikan_sarana')->insert( $records );
 		}
-
-		DB::table('app_kepemilikan_sarana')->insert( $records );
 
 		$r->session()->flash('success','Data tersimpan');
